# Add new `Performance/BlockGivenWithExplicitBlock` cop