@fatkodima 2020-10-02
Good point. Updated to not trigger an offense when the block arg is reassigned.